Refactor creating salt.loader.lazy
